paligemma896/pyproject.toml
[project]
name = "paligemma896"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.1.1",
    "pdf2image>=1.17.0",
    "pillow>=11.0.0",
    "requests>=2.32.3",
    "torch>=2.5.1",
    "transformers>=4.46.3",
]
